Better handle error conditions
When ScanTask encounters an error, log it instead of silently dying
and not processing any more refs.  This also exposed a few bugs in
JGit: there's an erroneous assumption that the submodule name will
be the same as the path, and the UTF-8 BOM is not handled cleanly by
BlobBasedConfig, which results in an error from SubmoduleWalk.

Change-Id: I280204390ab5fcf60dd6d0432c8acb17d85e3cb8